Rename Resonate to Sendspin
