--- music_assistant/providers/resonate/manifest.json
@@ -1,9 +1,0 @@
-{
-  "type": "player",
-  "domain": "resonate",
-  "stage": "alpha",
-  "name": "Resonate (WIP)",
-  "description": "Resonate (working title) is the next generation streaming protocol built by the Open Home Foundation. Follow the development on Discord to see how you can get involved.",
-  "codeowners": ["@music-assistant"],
-  "requirements": ["aioresonate @ git+https://github.com/maximmaxim345/aioresonate.git@feat/update-spec"]
-}
